Move libvpx/* to *
This is needed to get the folder structure to be in sync with main
branch of upstream project.

Except the changes in Android.bp and generate_config.sh
(as per the new folder structure), all other changes in this CL
are due to files being moved without any changes in the code.

Bug: 256050947
Test: Builds

Change-Id: Ic62f7b1bd28cc0d2d84c8cb8c6a0062ac7fb68af